backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
aiomysql>=0.2.0
redis>=5.0.0
aiokafka>=0.11.0
motor>=3.6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyJWT>=2.8.0
httpx>=0.27.0
websockets>=13.0
python-dotenv>=1.0.0
